backend/requirements.txt
Django
djangorestframework
django-cors-headers
django-environ
channels
PyMuPDF
requests
tree-sitter==0.21.3